chore: create separated main and use FludemyApp
