+++ lib/app/pages/home/widgets/appbar/web_app_bar_responsive_content.dart
