+++ lib/app/pages/home/widgets/sections/course_item.dart
